Update packages
This also adds an explicit override to react-scripts due to a
vulnerability in a transient package.

See issue in [create-react-app](https://github.com/facebook/create-react-app/issues/11174)
and linked open PR to bump the dependency [here](https://github.com/facebook/create-react-app/pull/12172).

For now, we are adding our override to bump the version of webpack.