Add ability to expand story into detail view
Some changes needed to be made in order to preserve the filtered story
points so we could filter the map down to just story places and then
refresh the map back to all filtered stories when the place detail is
closed.

We also needed to adjust how to stored and managed the filters so the
filter select state did not get trashed when hiding/showing the filters.