fix: Community Switcher doesn't update map configuration
Before, community switcher would keep the prior (or first) loaded map
settings (projection, zoom, bounds, etc). The mapcontext attempted to
update this, but because most of that it set once and left alone (as it
should), the switcher and map configuration couldn't update with the
selected community's settings.

Now, the Map has the configuration passed to the component (rather than
saved in context) so it always updates as expected and further, we added
the community's slug as the key to the map so it forces a full reload
when the switch occurs. This ensures that any open popups are also
removed, among all of the other oddities we were encountering.