fix(temp): remove placename audio player from detail card
When media (particularly placenameAudio) is present on a Place and the
Place chip was visisble, clicking on a Story in that place's story list
would cause it to catastrophically fail. This was due to react trying to
remove the Media (I guess?) but the node wasn't available to remove (for
some reason). I tried to fix it properly, but there wasn't an easy
solution that didn't require a significant amount of refactoring or
breaking the ability to breadcrumb through place > story and then
"close" the story and see the place again.

For now, I've removed the code to allow the place audio name to be
played from the chip altogether. Since it's available in the popup on
the map view, this seems okay for now until we get more user feedback.