fix: display loading rather than error when pending mount
ReactRouter uses fallbackElement to display to show the user that the
site is "working" but still pending a full mount. We were incorrectly
showing the "NotFound" page, which resulted in occassionally flashing
the error screen when nothing went wrong. We now correctly show the
loading spinner when things are pending.