fix: disable revalidation on data loader parent
data loader routers will attempt to revalidate on specific triggers.
This is usually expected, and desired, but in our case, it caused double
renders.

This is because we use a "parent" route with a rendered Layout element.
Since our homeLoader is on a child nested route, but marked as "index:
true" so it's treated as root, when that child revalidated, so did the
parent.

The solve is to just never allow the parent element to revalidate. Each
child route can manage their own revalidation; but the Layout parent
node should never need to.