fix: router instantiating routes twice
This moves the router declaration outside of the Component so when
React.StrictMode renders the component it twice, it won't instantiate
the router twice. This ensures we are not double rendering the router
when App needs to load again (aka: root load on first hit or when
removing all query params).

This also cleans up the multiple uses of React.Suspense that didn't do
anything, and moves all of the Sidebar search functionality into the
Sidebar component rather than passing props. And fixes a few areas where
we were running startTransition when no transition needed to occur (thus
forcing a second render).