fix: don't fire window resize handler when entering fullscreen
Full screen mode adjusts the window size; and while we typically want to
handle all resize events, we need to ignore resizing when entering
fullscreen.

If we allow it to resize, when the browser window is smaller than the
fullwidth of the device screen, entering full screen would change the
width which changed isMobile which immediately rerenders the component
causing fullscreen mode to immediately close.