fix: long place descriptions make sidebar unscrollable
Super long descriptions on a Place would make the place chip card expand
greater than length of screen. This makes scrolling to see the places
stories impossible.

This adds a container for the description and forces the height to 30vh
of the device and auto-scrolls long descriptions.

We may prefer a hide/read more in the future, but this was a quick win.