enhancement: allow fitBounds to be configured per request
Before, the map would always update to "fit" all of the points into the
map view — even when that could disrupt how the user was interacting
with that map.

Now, this can be configured any time we tell the map to update points.
When updateBounds is true, we calculate map bounds and center using the
points in the featureCollection. When false, we zoom the map to the
configured zoom level for the community.

Specifically,
1. When filters are cleared, we zoom.
2. When options are cleared or the last option is removed, we zoom.
3. When an option is removed and there is still at lesat one option, we
   fit.
4. When a Story is clicked, we fit.
5. When a Place is clicked, we fit.
6. When a Place is cleared and prior points were stashed, we fit.
7. When a Place is cleared and no points are stashed, we zoom.
8. When a Story is cleared and a Place is selected, we fit to Place.
9. When a Story is cleared and no place is selected, we zoom.

This matches more closely with the main Terrastories map behavior and
aligns with customer requests that we've had about the map zooming away
from the region being explored.