fix: overscroll behavior
Before we were disabling overscroll behavior when we started dragging
the sidepanel. This had an odd behavior in Safari where it would
"flicker" the bounce effect and occassionaly still trigger the
overscroll pull to refresh feature.

Now, we are disabling when the bar is first touched/tapped and enabling
when released; regardless of move state. This fixes the wonky behavior
on Safari without modifying behavior in other browsers.