fix(mobile): calculate pageHeight when needed
Before we were setting pageHeight to window.innerHeight once when the
function was loaded. While this worked most of the time, occassionally
the height calculation for where to open the panel would result in
getting pulled visually above the top of the page.

Since reproducing is difficult and sharing map moving state with the
side panel, this will hopefully fix the issue without needing a more
robust solution.