feature: New Story list layout
This changes the layout and information displayed in the Story list when
exploring a community map. Before, we showed text-based title and
description which can cause the list to look sparse and uninteresting to
navigate.

Now, when available, we show a media preview thumbnail w/ media type
indicators overlayed. When a thumbnail is not available, but there is
media available, the first type is shown as an Icon as a thumbnail
placeholder with any remaining icons overlayed. And finally, if there
are no media indicators, we show a grayed out Terrastories logomark.

Furthermore, rather than showing the description of a Story, we show
just the Title and the Topic in a badge.

All of these create a more visually appealing list that is still easy to
use, but less text heavy and more visual.