enhancement: Fix isMobile toggleOverscroll func
This updates and enhances our toggleOverscroll behavior. Previously, we
were running `toggle` on the classList to add/remove as the reverse of
what was previously there.

Now, it's explicitly toggled based on a conditionally. This allows us to
ensure that if we are calling toggle twice, we aren't accidentally
enabling and immediately disabling (or the reverse). In addition, we've
updated the function name and call to reflect these changes.